Refactor packaging for clarity
