+++ ide/src/main/java/com/chiralbehaviors/graphql/layout/controls/NestedTableRow.java
@@ -18,7 +18,7 @@
  *  along with Ultrastructure.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.chiralbehaviors.graphql.layout;
+package com.chiralbehaviors.graphql.layout.controls;
 
 import java.util.Arrays;
 import java.util.HashMap;
