+++ ide/src/main/java/com/chiralbehaviors/graphql/layout/explorer/QueryState.java
@@ -18,7 +18,7 @@
  *  along with Ultrastructure.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.chiralbehaviors.graphql.layout;
+package com.chiralbehaviors.graphql.layout.explorer;
 
 public class QueryState {
     private String data;
