+++ ide/src/main/resources/com/chiralbehaviors/graphql/layout/graphiql.fxml
@@ -10,7 +10,7 @@
 
 <BorderPane prefHeight="400.0" prefWidth="600.0"
 	xmlns="http://javafx.com/javafx/8.0.65" xmlns:fx="http://javafx.com/fxml/1"
-	fx:controller="com.chiralbehaviors.graphql.layout.GraphiqlController">
+	fx:controller="com.chiralbehaviors.graphql.layout.explorer.GraphiqlController">
 	<top>
 		<GridPane BorderPane.alignment="CENTER">
 			<columnConstraints>
